[refactor] 데이터 불러오는 동안 loader 구현
